requirements.txt
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
psycopg2-binary==2.9.7
python-decouple==3.8
django-cors-headers==4.3.1
Pillow==10.0.1
boto3==1.29.7
django-storages==1.14.2
stripe==7.8.0
python-magic==0.4.27
opencv-python==4.8.1.78
celery==5.3.4
redis==5.0.1
reportlab==4.0.7
psutil==5.9.6